feat(user): add recipes and subscribers count to user profile
This commit adds two new fields to the `User` type in the GraphQL schema:

1. `recipesCount`: an integer representing the number of recipes the user has created.
2. `subscribersCount`: an integer representing the number of users who have subscribed to the user.

These fields are marked with the `@cost(weight: "10")` directive, indicating that they have a significant performance impact and should be used judiciously.

The corresponding changes have been made in the frontend `CookCard` component to display these new fields.