feat(tag): add ability to create new tags
This commit adds the ability to create new tags in the TagSelector component of the frontend application. It introduces a new mutation called `AddTagMutation` that allows users to create a new tag by providing a name. The backend has also been updated to include an `Authorize` attribute on the `UpdateTag` and `DeleteTagById` mutations to ensure that only authenticated users can perform these actions.